+++ src/main/java/net/silvertide/kindred/Kindred.java
@@ -4,9 +4,11 @@ import com.mojang.logging.LogUtils;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.silvertide.kindred.config.ClientConfig;
 import net.silvertide.kindred.config.Config;
 import net.silvertide.kindred.network.Networking;
+import net.silvertide.kindred.registry.ModAttributes;
 import org.slf4j.Logger;
 
 @Mod(Kindred.MODID)
@@ -16,6 +18,7 @@ public class Kindred {
 
     public Kindred() {
         Networking.register();
+        ModAttributes.register(FMLJavaModLoadingContext.get().getModEventBus());
         ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, Config.SPEC);
         ModLoadingContext.get().registerConfig(ModConfig.Type.CLIENT, ClientConfig.SPEC);
     }