+++ src/main/java/net/silvertide/kindred/compat/pmmo/PmmoBridge.java
@@ -1,7 +1,0 @@
-package net.silvertide.kindred.compat.pmmo;
-
-import net.minecraft.world.entity.player.Player;
-
-public interface PmmoBridge {
-    long getSkillLevel(Player player, String skill);
-}